requirements.txt
Django==1.6.4
MySQL-python==1.2.5
South==0.8.4
django-allauth==0.12.0
django-tastypie==0.11.0
oauthlib==0.6.1
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
wsgiref==0.1.2
xlrd==0.9.3
yolk==0.4.3
